+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -135,7 +135,8 @@ final class AndroidOptionsInitializer {
     options.setEnvelopeDiskCache(new AndroidEnvelopeCache(options));
 
     final ActivityFramesTracker activityFramesTracker =
-        new ActivityFramesTracker(loadClass, options.getLogger());
+        new ActivityFramesTracker(loadClass, options);
+
     installDefaultIntegrations(
         context,
         options,